# packages/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/documents": {
      "get": {
        "operationId": "DocumentsController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "post": {
        "operationId": "DocumentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/upload": {
      "post": {
        "operationId": "DocumentsController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}/download": {
      "get": {
        "operationId": "DocumentsController_download",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}/reupload": {
      "post": {
        "operationId": "DocumentsController_reuploadFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReuploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}/versions": {
      "get": {
        "operationId": "DocumentsController_findVersions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
    "/documents/{id}": {
      "get": {
        "operationId": "DocumentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      },
      "patch": {
        "operationId": "DocumentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Documents"
        ]
      }
    },
...
    "/external-references": {
      "get": {
        "operationId": "ExternalReferencesController_findAllForIssue",
        "parameters": [
          {
            "name": "issueId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ExternalReferences"
        ]
      },
      "post": {
        "operationId": "ExternalReferencesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExternalReferenceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ExternalReferences"
        ]
      }
    },
    "/external-references/{id}": {
      "delete": {
        "operationId": "ExternalReferencesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ExternalReferences"
        ]
      }
    }
  },
  "info": {
    "title": "EPG Platform API",
    "description": "Enterprise Project Governance Platform — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {}
      },
      "UploadDocumentDto": {
        "type": "object",
        "properties": {}
      },
      "ReuploadDocumentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateDocumentDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateExternalReferenceDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}